// src/joi-x-validators.d.ts
export declare enum AuthType {
    password = "password",
    publicKey = "publicKey",
    any = "any",
}
export declare enum PassType {
    plainText = "plainText",
    encrypt = "encrypt",
}
export declare type AuthTypes = AuthPassword | AuthPublicKey | AuthAny;
export declare type AuthPassword = JoiX.ExtractFromSchema<typeof authPassword>;
export declare type AuthPublicKey = JoiX.ExtractFromSchema<typeof authPublicKey>;
export declare type AuthAny = JoiX.ExtractFromSchema<typeof authAny>;
export declare const password: (passType?: PassType) => JoiX.XObject & JoiX.ObjectSchema & {
    __tsTypeO: {
        phrase: JoiX.XPrimitive<string> & JoiX.StringSchema & {
            __isRequired: "T";
        };
        type: JoiX.XStringSchema<PassType>;
    };
};